update
• Updated input 'hyprland':
    'git+https://github.com/hyprwm/Hyprland?ref=refs/heads/main&rev=1cc35cfd817ed5dc87cb3e50e2bd46b87cf17c33' (2024-06-02)
  → 'git+https://github.com/hyprwm/Hyprland?ref=refs/heads/main&rev=358e59e69d27a69381bc0872b5b8d1184bc6b6a1' (2024-06-02)
• Updated input 'mypkgs':
    'github:spitulax/mypkgs/04c5ef2b1a5ef367371592ff50d0d4c537169146' (2024-06-03)
  → 'github:spitulax/mypkgs/5191b475850e1f8ae51cf1ba7dd5580e9c7f424c' (2024-06-03)
• Updated input 'mypkgs/nixpkgs':
    'github:nixos/nixpkgs/ad57eef4ef0659193044870c731987a6df5cf56b' (2024-05-29)
  → 'github:nixos/nixpkgs/57610d2f8f0937f39dbd72251e9614b1561942d8' (2024-05-31)
• Updated input 'nixpkgs-unstable':
    'github:nixos/nixpkgs/ad57eef4ef0659193044870c731987a6df5cf56b' (2024-05-29)
  → 'github:nixos/nixpkgs/57610d2f8f0937f39dbd72251e9614b1561942d8' (2024-05-31)

+++ packages/scripts/lazyup.nix
@@ -2,10 +2,9 @@
 writeShellScriptBin "lazyup" ''
 
 LAZYLOCK_PATH=$HOME/.config/nvim/lazy-lock.json
-[[ ! -r $LAZYLOCK_PATH ]] && echo "$LAZYLOCK_PATH does not exist" && exit 1
-
 cd $FLAKE
-unlink $LAZYLOCK_PATH
+[[ -r $LAZYLOCK_PATH ]] && unlink $LAZYLOCK_PATH
+
 nvim . # run lazy.nvim commands
 cp $LAZYLOCK_PATH modules/home/bintang/nvim/lazy-lock.json
 rm $LAZYLOCK_PATH
